requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.13
django-allauth==0.39.1
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-taggit==0.23.0
django-bootstrap4==1.1.1
django-widget-tweaks==1.4.11
idna==2.8
jsonfield==2.0.2
Markdown==2.6.11
oauthlib==3.0.1
Pillow==5.3.0
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.3

